move topic service and s3 client to service.topics
